# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: POS Restaurant API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tables:
    get:
      operationId: listTables
      tags: [tables]
      summary: List all tables
      responses:
        "200":
          description: List of tables
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Table"
    post:
      operationId: createTable
      tags: [tables]
      summary: Create a table
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTableBody"
      responses:
        "201":
          description: Created table
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Table"
  /tables/{id}:
    get:
      operationId: getTable
      tags: [tables]
      summary: Get table by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Table
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Table"
        "404":
          description: Not found
    patch:
      operationId: updateTable
      tags: [tables]
      summary: Update table
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTableBody"
      responses:
        "200":
          description: Updated table
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Table"
    delete:
      operationId: deleteTable
      tags: [tables]
      summary: Delete table
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Orders
...
components:
  schemas:
...
    Table:
      type: object
      properties:
        id:
          type: integer
        number:
          type: integer
        name:
          type: string
        seats:
          type: integer
        status:
          type: string
          enum: ["free", "occupied", "reserved"]
        createdAt:
          type: string
      required:
        - id
        - number
        - name
        - seats
        - status
        - createdAt
    CreateTableBody:
      type: object
      properties:
        number:
          type: integer
        name:
          type: string
        seats:
          type: integer
        status:
          type: string
          enum: ["free", "occupied", "reserved"]
      required:
        - number
        - name
        - seats
        - status
    UpdateTableBody:
      type: object
      properties:
        number:
          type: integer
        name:
          type: string
        seats:
          type: integer
        status:
          type: string
          enum: ["free", "occupied", "reserved"]
...